[update] update AuthServiceImpl, AuthController

+++ src/main/java/com/example/project/auth/controller/AuthController.java
@@ -3,6 +3,7 @@ package com.example.project.auth.controller;
 import com.example.project.auth.exception.TypeNotFound;
 import com.example.project.auth.requestbody.CreateAuthRequest;
 import com.example.project.auth.requestbody.UpdateAuthRequest;
+import com.example.project.auth.requestbody.UpdateLoginRequest;
 import com.example.project.auth.service.AuthService;
 import com.example.project.auth.view.AuthView;
 import io.swagger.annotations.ApiImplicitParam;
@@ -46,15 +47,32 @@ public class AuthController {
 
     @PutMapping("")
     public ResponseEntity<?> updateAuthInfo(@RequestBody @Valid UpdateAuthRequest updateAuthRequest, HttpServletRequest request) {
+
+//        if (updateAuthRequest.getType().equals("회원정보 수정")) {
+//            authService.updateAuthInfo(updateAuthRequest, request);
+//            return ResponseEntity.ok("회원정보가 수정되었습니다.");
+//
+//        } if (authService.deleteAuthInfo(updateAuthRequest, request)) {
+//            if (updateAuthRequest.getType().equals("회원 탈퇴")) {
+//                authService.deleteAuthInfo(request, updateAuthRequest);
+//                return ResponseEntity.ok("회원탈퇴가 되었습니다.");
+//            }
+//        }
+
+        Boolean auth = authService.deleteAuthInfo(request, updateAuthRequest);
+
         if (updateAuthRequest.getType().equals("회원정보 수정")) {
             authService.updateAuthInfo(updateAuthRequest, request);
             return ResponseEntity.ok("회원정보가 수정되었습니다.");
-
-        } else if (updateAuthRequest.getType().equals("회원 탈퇴")) {
-            authService.deleteAuthInfo(request, updateAuthRequest);
-            return ResponseEntity.ok("회원탈퇴가 되었습니다.");
         }
-        throw new TypeNotFound();
+        if (updateAuthRequest.getType().equals("회원 탈퇴")) {
+            if (auth != null) {
+                authService.deleteAuthInfo(request, updateAuthRequest);
+                return ResponseEntity.ok("회원탈퇴가 되었습니다.");
+            }
+            throw new TypeNotFound();
+        }
+        return null;
     }
 
     @GetMapping("")
